Omry/compiler ast state (#35)
* fixed typeo

* changed cClass defaults

* refactor+fixes ast based compiler

* fixed failing tests duew to new eqlCode matcher limitations

* fixed some more issues with ast compiler

* all but $$processupdate looking good

* moved the is funcs to their own file

* updated expected js files

* added more stateful functionality

* kinda broken

* fixed testing issue: child text matcher

* fixed string template compiler, to work with props.xxx

* removed log

* fixed dom binding

* state working in AST compiler, event handlers not so much

* trimCode fix

* minor fix

* fixed example3 to new framework assumptions

* added feature based testing of compilers

* minor issues fix

* fixed playground compileExample issue

* added progressive tollerence to time based tests

* fixed utils build (removed __filename)

* updated manually compiled tests

* added missing semi

* updated tests

* laid some groundwork to handle events

* added toString support for evemt handlers

* added event handlers

* added event handler

* added props support

* improved binding support

* compiles but had runtime errors

* adding sourcemaps

* fixed event handlers

* getting closer to state compilation

* have a nice reporter going

* added reported tolerance to db failure

* reporter knida ok

* state change working, side effected not

* updated examples and improved func mod

* ast looking good

* fixed prepputeer retry bug

* updated manually compiled

* fixed string compiler

* minor test fix

* fixed lint/tsc errors

* added some more time for slow tests

* added some more testing time

* oops forgot to check the new feature

* windows fix

* improved logging

* improved error message

* added src.tmp to gitignore

* improved reporting

* llint fixes

* typeo

* breaking down generators

* fixed broken imports

* added aggressive git clean to push task

* cleanups

* fixed browserify tests in windows

+++ packages/testing/src/test.using.examples.ts
@@ -1,16 +1,16 @@
 import { after } from 'mocha';
-import { Compiler, ExamplePaths } from '@tsx-air/types';
+import { Compiler, ExamplePaths, Features, ALL } from '@tsx-air/types';
 import { loadSuite } from '@tsx-air/testing';
-import ts from 'typescript';
 import { browserify } from '@tsx-air/browserify';
 import { join, basename } from 'path';
 import rimraf from 'rimraf';
 import { preppeteer } from './html/puppeteer.mocha.utils';
 import { packagePath } from '@tsx-air/utils/packages';
+import { safely } from '@tsx-air/utils/src';
 
 const fixtures = packagePath('@tsx-air/examples', 'fixtures');
 const publicPath = packagePath('@tsx-air/examples', 'public');
-const tempPath = packagePath('@tsx-air/examples', '.tmp');
+const tempPath = packagePath('@tsx-air/examples', 'tmp');
 
 export function shouldCompileExamples(compiler: Compiler, examplePaths: string[]) {
     const examples = examplePaths.map(loadSuite);
@@ -21,43 +21,69 @@ export function shouldCompileExamples(compiler: Compiler, examplePaths: string[]
         });
 
         examples.map(
-            ({ suite, path }) => {
+            ({ suite, path, features }) => {
                 const exampleName = basename(path);
                 const paths: ExamplePaths = {
                     temp: join(tempPath, exampleName, Date.now().toString(36)),
                     fixtures,
                     path
                 };
-
-                describe(exampleName, () => {
-                    before(() => {
-                        this.timeout(process.env.CI ? 15000 : 6000);
-                        return browserifyBoilerplate(path, paths.temp, compiler.transformers);
+                const unsupported = getUnsupported(features, compiler);
+                if (unsupported.length) {
+                    describe(exampleName, () => {
+                        it.skip(`Unsupported features\n${
+                            unsupported.map(f => '\t\t' + [...f.values()].join(' ')).join('\n')}`, () => {
+                                /* */
+                            });
                     });
-                    beforeEach(() => Promise.all([
-                        api.server.addStaticRoot(path),
-                        api.server.addStaticRoot(paths.temp),
-                    ]));
-                    after(function () {
-                        if (this.test?.parent?.tests.every(t => t.isPassed())) {
-                            rimraf(paths.temp, () => null);
-                        }
-                    });
+                } else {
+                    describe(exampleName, () => {
+                        before(async () => {
+                            this.timeout(process.env.CI ? 15000 : 6000);
+                            this.retries(0);
+                            return safely(
+                                () => browserifyBoilerplate(path, paths.temp, compiler),
+                                'Failed to compile'
+                            );
+                        });
+                        beforeEach(() => Promise.all([
+                            api.server.addStaticRoot(path),
+                            api.server.addStaticRoot(paths.temp),
+                        ]));
+                        after(function () {
+                            if (this.test?.parent?.tests.every(t => t.isPassed())) {
+                                rimraf(paths.temp, () => null);
+                            }
+                        });
 
-                    suite.call(this, api, paths);
-                });
+                        suite.call(this, api, paths);
+                    });
+                }
             });
     });
 }
 
+const getUnsupported = (features: Features, compiler: Compiler) => {
+    if (compiler.features === ALL) {
+        return [];
+    }
+    return features.filter(tested => !compiler.features.find(
+        supported => {
+            for (const subFeature of tested) {
+                if (!supported.has(subFeature)) {
+                    return false;
+                }
+            }
+            return true;
+        }
+    ));
+};
+
 const browserifyBoilerplate = async (examplePath: string, target: string,
-    transformers: ts.CustomTransformers) => await browserify({
+    compiler: Compiler) => await browserify({
         base: examplePath,
         entry: 'suite.boilerplate.ts',
         output: join(target, 'boilerplate.js'),
         debug: !!process.env.DEBUG,
-        loaderOptions: {
-            transformers,
-            cache: false
-        }
+        compiler
     });
